backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
moviepy==1.0.3
sqlalchemy
pymysql
ffmpeg-python
faster-whisper
requests